backend/requirements.txt
fastapi
uvicorn
python-multipart
langchain
langchain-community
langchain-google-genai
pypdf
chromadb
pydantic
pydantic-settings
python-dotenv
sqlalchemy
psycopg2-binary
